lib/rst/interp_float clean-up:  Use off_t for offsets  Fully type function pointers  Fix type of check_points field in struct interp_params to match   IL_check_at_points_2d()
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@38715 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/rst/interp_float/interp2d.c
@@ -45,7 +45,7 @@ int IL_grid_calc_2d(struct interp_params *params, struct quaddata *data,	/* give
 		    double *c1min, double *c1max, double *c2min, double *c2max,	/* min and max interp. curv. val. */
 		    double *ertot,	/* total interplating func. error */
 		    double *b,	/* solutions of linear equations */
-		    int offset1,	/* offset for temp file writing */
+		    off_t offset1,	/* offset for temp file writing */
 		    double dnorm)
 
 /*
@@ -80,7 +80,7 @@ int IL_grid_calc_2d(struct interp_params *params, struct quaddata *data,	/* give
     double zz;
     int bmask = 1;
     static int first_time_z = 1;
-    int offset, offset2;
+    off_t offset, offset2;
     double fstar2 = params->fi * params->fi / 4.;
     double tfsta2, tfstad;
     double ns_res, ew_res;
